requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
google-cloud-firestore>=2.11.0
firebase-admin>=6.2.0
pydantic>=1.10.7
python-dotenv>=1.0.0
email-validator>=2.0.0
aiosmtplib>=2.0.0
jinja2>=3.1.6
cloudinary>=1.36.0
redis>=5.0.0
pytz>=2023.3

starlette~=0.46.2
python-multipart>=0.0.20